change distance to score
