+++ packages/toolbox_store/src/toolbox_store/db.py
@@ -7,11 +7,12 @@ import sqlite_vec
 
 from toolbox_store.filters import build_where_clause
 from toolbox_store.models import (
+    RetrievedChunk,
     StoreConfig,
     TBDocument,
+    TBDocumentChunk,
     is_valid_field_identifier,
 )
-from toolbox_store.retrieved_chunks import RetrievedChunk, TBDocumentChunk
 
 T = TypeVar("T", bound=TBDocument)
 
@@ -375,6 +376,11 @@ class TBDatabase(Generic[T]):
                 row_dict["embedding"] = deserialize_float32(embedding_blob)
             else:
                 row_dict["embedding"] = None
+
+            # Convert distance to score
+            distance = row_dict.get("distance", None)
+            if distance is not None:
+                row_dict["score"] = 1 - distance
 
             results.append(RetrievedChunk.from_sql_row(row_dict))
