+++ packages/toolbox_store/src/toolbox_store/reranking.py
@@ -51,10 +51,63 @@ def rerank_chunks(
     for res in reranked.results:
         doc_id = res.document.metadata["document_id"]
         chunk_idx = res.document.metadata["chunk_idx"]
-        score = res.score if res.score is not None else 0.0
-        distance = -score
         chunk = chunks_by_id[(doc_id, chunk_idx)]
-        chunk.distance = distance
+        chunk.score = res.score if res.score is not None else -1
         reranked_chunks.append(chunk)
 
     return reranked_chunks
+
+
+# def _get_normalize_bounds(
+#     reranker: BaseRanker,
+#     results: list[RetrievedChunk],
+#     query: str,
+# ) -> tuple[float, float]:
+#     min_score = np.inf
+#     max_score = -np.inf
+#     all_scores = [-res.distance for res in results if res.distance is not None]
+#     if all_scores:
+#         min_score = min(all_scores)
+#         max_score = max(all_scores)
+
+#     bounds_ranked = reranker.rank(
+#         query=query,
+#         docs=[DUMMY_DOC, query],
+#         doc_ids=[0, 1],
+#     ).results
+
+#     bounds = [min_score, max_score]
+#     for res in bounds_ranked:
+#         if res.score is not None:
+#             bounds[0] = min(bounds[0], res.score)
+#             bounds[1] = max(bounds[1], res.score)
+
+#     return bounds[0], bounds[1]
+
+
+# def _normalize(
+#     chunks: list[RetrievedChunk], min_val: float = 0.0, max_val: float = 1.0
+# ) -> list[RetrievedChunk]:
+#     """Normalize chunk distances to a specified range [min_val, max_val]."""
+#     if not chunks:
+#         return chunks
+
+#     distances = [chunk.distance for chunk in chunks if chunk.distance is not None]
+#     if not distances:
+#         return chunks
+
+#     min_score, max_score = min(distances), max(distances)
+
+#     if min_distance == max_distance:
+#         for chunk in chunks:
+#             chunk.distance = (min_val + max_val) / 2.0
+#         return chunks
+
+#     for chunk in chunks:
+#         if chunk.distance is not None:
+#             normalized_score = (chunk.distance - min_distance) / (
+#                 max_distance - min_distance
+#             )
+#             chunk.distance = normalized_score * (max_val - min_val) + min_val
+
+#     return chunks
